# cmd/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.Review:
    properties:
      animeId:
        description: ID of the anime being reviewed
        example: 1
        type: integer
      content:
        description: Content of the review
        example: This anime is amazing!
        maxLength: 500
        type: string
      created_at:
        description: Timestamp when the review was created
        example: "2025-02-20T19:27:00Z"
        type: string
      id:
        description: Unique identifier for the review
        example: 1
        type: integer
      rating:
        description: Rating given in the review (0-10)
        example: 9
        maximum: 10
        minimum: 0
        type: integer
      updated_at:
        description: Timestamp when the review was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      userId:
        description: ID of the user who created the review
        example: 1
        type: integer
    required:
    - content
    - rating
    type: object
...
  models.User:
    properties:
      created_at:
        description: Timestamp when the user was created
        example: "2025-02-20T19:27:00Z"
        type: string
      email:
        description: Email address of the user (omitted unless necessary)
        example: john@example.com
        maxLength: 100
        minLength: 5
        type: string
      id:
        description: Unique identifier for the user
        example: 1
        type: integer
      is_admin:
        description: Indicates if the user has admin privileges
        example: false
        type: boolean
      reviews:
        description: List of reviews created by the user (omitted unless necessary)
        items:
          $ref: '#/definitions/models.Review'
        type: array
      updated_at:
        description: Timestamp when the user was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      username:
        description: Unique username for the user
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
  models.UserCreateRequest:
    properties:
      email:
        description: Email address for the new user
        example: john@example.com
        maxLength: 100
        minLength: 5
        type: string
      password:
        description: Password for the new user
        example: password123
        maxLength: 100
        minLength: 5
        type: string
      username:
        description: Username for the new user
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        description: Timestamp when the user was created
        example: "2025-02-20T19:27:00Z"
        type: string
      id:
        description: Unique identifier for the user
        example: 1
        type: integer
      is_admin:
        description: Indicates if the user has admin privileges
        example: false
        type: boolean
      updated_at:
        description: Timestamp when the user was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      username:
        description: Username of the user
        example: john_doe
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@myanimeapi.com
    name: API Support
  description: This is a sample API for managing anime and reviews.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: MyAnimeAPI
  version: "1.0"
paths:
...
  /v1/users:
    get:
      description: Retrieve a paginated list of all users. Requires admin privileges.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Invalid pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve users
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get all users (admin only)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided data.
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid input or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: User with this username or email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create user
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - "":
        - ""
      summary: Create a new user
      tags:
      - users
  /v1/users/{id}:
    delete:
      description: Delete a user by their ID. Requires authentication.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete user
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      description: Retrieve a user by their ID. Requires authentication.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve user
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user with the provided data. Requires authentication.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid input or ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update user
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Use the format "Bearer <JWT_TOKEN>". Example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"